add game

+++ code/inc/EC_Core/IGameDesc.h
@@ -1,6 +1,7 @@
 #ifndef _EASY_CARD_GAME_DESCRIPTOTION_INTERFACE_H_
 #define _EASY_CARD_GAME_DESCRIPTOTION_INTERFACE_H_
 
+#include "EC_Common/type_def.h"
 
 namespace EasyCard
 {
@@ -9,7 +10,7 @@ namespace EasyCard
 	public:
 		virtual const char* GetName() = 0;
 		virtual const char* GetDisplayName() = 0;
-		virtual int GetPlayerNumber() = 0;
+		virtual size_t GetPlayerNumber() = 0;
 	};
 }
 #endif
